Finished Playlist support

+++ WinSonic/Pages/Details/PlaylistDetailPage.xaml.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
+using System.Linq;
 using WinSonic.Controls;
 using WinSonic.Model.Api;
 using WinSonic.Model.Player;
@@ -26,15 +27,17 @@ namespace WinSonic.Pages.Details
         private void OnPropertyChanged(string propertyName) =>
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         public InfoWithPicture? DetailedObject { get; set; }
-        public ObservableCollection<Song> Songs { get; set; } = [];
+        public DetailedPlaylistAdapter Playlist { get; private set; }
         private readonly App app = (App)Application.Current;
         private Song? RightClickedSong;
         private CommandBarFlyout? _songFlyout;
+        private bool IsOwner;
 
         public PlaylistDetailPage()
         {
             InitializeComponent();
             SongGridTable.Columns = [
+                new Tuple<string, GridLength>("Track", new GridLength(80, GridUnitType.Pixel)),
                 new Tuple<string, GridLength>("Title", new GridLength(4, GridUnitType.Star)),
                 new Tuple<string, GridLength>("Artist", new GridLength(3, GridUnitType.Star)),
                 new Tuple<string, GridLength>("Album", new GridLength(3, GridUnitType.Star)),
@@ -55,27 +58,24 @@ namespace WinSonic.Pages.Details
         {
             if (DetailedObject != null)
             {
-                var playlist = await SubsonicApiHelper.GetPlaylist(DetailedObject.ApiObject.Server, DetailedObject.ApiObject.Id);
-                if (playlist != null)
-                {
-                    NoteTextBlock.Text = playlist.Comment;
-                    foreach (var song in playlist.Entry)
-                    {
-                        Songs.Add(new Song(song, DetailedObject.ApiObject.Server));
-                    }
-                }
-
-                foreach (var song in Songs)
+                Playlist = new(await SubsonicApiHelper.GetPlaylist(DetailedObject.ApiObject.Server, DetailedObject.ApiObject.Id));
+                OnPropertyChanged(nameof(Playlist));
+                IsOwner = string.Equals(Playlist.Playlist.Owner, Playlist.Playlist.Server.Username, StringComparison.OrdinalIgnoreCase);
+                OnPropertyChanged(nameof(IsOwner));
+                int i = 1;
+                foreach (var song in Playlist.Songs)
                 {
                     TimeSpan duration = TimeSpan.FromSeconds(song.Duration);
                     Dictionary<string, string?> dic = new()
                     {
+                        ["Track"] = string.Format("{0:D" + Playlist.Songs.Count.ToString().Length + "}", i),
                         ["Title"] = song.Title,
                         ["Artist"] = song.Artist,
                         ["Album"] = song.Album,
                         ["Time"] = string.Format("{0:D1}:{1:D2}", duration.Minutes, duration.Seconds),
                     };
                     SongGridTable.AddRow(dic);
+                    i++;
                 }
                 SongGridTable.ShowContent();
             }
@@ -89,7 +89,7 @@ namespace WinSonic.Pages.Details
 
         private void AddToQueueButton_Click(object sender, RoutedEventArgs e)
         {
-            foreach (var song in Songs)
+            foreach (var song in Playlist.Songs)
             {
                 PlayerPlaylist.Instance.AddSong(song);
             }
@@ -97,21 +97,21 @@ namespace WinSonic.Pages.Details
 
         private void PlayNextButton_Click(object sender, RoutedEventArgs e)
         {
-            for (int i = Songs.Count - 1; i >= 0; i--)
+            for (int i = Playlist.Songs.Count - 1; i >= 0; i--)
             {
-                PlayerPlaylist.Instance.AddSong(Songs[i], (int)app.MediaPlaybackList.CurrentItemIndex + 1);
+                PlayerPlaylist.Instance.AddSong(Playlist.Songs[i], (int)app.MediaPlaybackList.CurrentItemIndex + 1);
             }
         }
 
         private void SongGridTable_RowDoubleTapped(object sender, RowEvent e)
         {
             PlayerPlaylist.Instance.ClearSongs();
-            PlayerPlaylist.Instance.AddSong(Songs[e.Index]);
+            PlayerPlaylist.Instance.AddSong(Playlist.Songs[e.Index]);
         }
 
         private CommandBarFlyout SongGridTable_RowRightTapped(object sender, RowEvent e)
         {
-            RightClickedSong = Songs[e.Index];
+            RightClickedSong = Playlist.Songs[e.Index];
             OnPropertyChanged(nameof(RightClickedSong));
             _songFlyout = SongCommandBarFlyout.Create(RightClickedSong, SongPlayButton_Click, SongPlayNextButton_Click, SongAddToQueueButton_Click, SongFavouriteButton_Click, SongAddToPlaylistButton_Click);
             return _songFlyout;
@@ -169,19 +169,55 @@ namespace WinSonic.Pages.Details
             }
         }
 
-        private void EditButton_Click(object sender, RoutedEventArgs e)
+        private async void EditButton_Click(object sender, RoutedEventArgs e)
         {
-
+            if (DetailedObject != null)
+            {
+                ContentDialog dialog = new()
+                {
+                    Title = "Editing playlist",
+                    Content = new UpdatePlaylistDialog(Playlist),
+                    PrimaryButtonText = "Update",
+                    CloseButtonText = "Cancel",
+                    XamlRoot = XamlRoot,
+                    Style = Application.Current.Resources["DefaultContentDialogStyle"] as Style,
+                    DefaultButton = ContentDialogButton.Primary
+                };
+                var result = await dialog.ShowAsync();
+                if (result == ContentDialogResult.Primary)
+                {
+                    Playlist.SaveChanges();
+                    await SubsonicApiHelper.UpdatePlaylist(Playlist.Playlist);
+                    app?.Window?.NavFrame.GoBack();
+                }
+                else
+                {
+                    Playlist.UndoChanges();
+                }
+            }
         }
 
         private async void DeleteButton_Click(object sender, RoutedEventArgs e)
         {
             if (DetailedObject != null)
             {
-                await SubsonicApiHelper.DeletePlaylist(DetailedObject.ApiObject.Server, DetailedObject.ApiObject.Id);
-                app?.Window?.NavFrame.GoBack();
+                ContentDialog dialog = new()
+                {
+                    Title = "Are you sure?",
+                    Content = "This action cannot be undone.",
+                    PrimaryButtonText = "Delete",
+                    CloseButtonText = "Cancel",
+                    XamlRoot = XamlRoot,
+                    Style = Application.Current.Resources["DefaultContentDialogStyle"] as Style,
+                    DefaultButton = ContentDialogButton.Close
+                };
+                var result = await dialog.ShowAsync();
+                if (result == ContentDialogResult.Primary)
+                {
+                    await SubsonicApiHelper.DeletePlaylist(DetailedObject.ApiObject.Server, DetailedObject.ApiObject.Id);
+                    app?.Window?.NavFrame.GoBack();
+                }
             }
-
         }
     }
 }
